vc4: Fix the argument type for cl_u16().
It doesn't matter, since it just got truncated to 16 inside, anyway.

+++ src/gallium/drivers/vc4/vc4_cl.h
@@ -55,7 +55,7 @@ cl_u8(struct vc4_cl *cl, uint8_t n)
 }
 
 static inline void
-cl_u16(struct vc4_cl *cl, uint32_t n)
+cl_u16(struct vc4_cl *cl, uint16_t n)
 {
         assert((cl->next - cl->base) + 2 <= cl->size);
